Fix build failures: update Hilt/Kotlin config and refactor MainActivity
Refactored `MainActivity.kt` to remove legacy AR rendering code and properly integrate `ProtractorScreen`.
Resolved Hilt compilation errors by providing `Gson` in `AppModule`.
Fixed Kotlin 2.3.0 compatibility issue by adding `kotlin-metadata-jvm` to KSP dependencies.
Fixed type mismatches in `MainLayout.kt` and `HaterScreen.kt`.
Recreated missing source files required for compilation.
Verified that `assembleDebug` and `testDebugUnitTest` pass.

+++ app/src/main/java/com/hereliesaz/cuedetat/ui/MainLayout.kt
@@ -63,9 +63,9 @@ fun MainLayout(
                 areHelpersVisible = uiState.areHelpersVisible,
                 experienceMode = uiState.experienceMode,
                 isTableVisible = uiState.table.isVisible,
-                tableSizeFeet = uiState.table.size.feet,
+                tableSizeFeet = uiState.table.size.feet.toFloat(),
                 isBeginnerViewLocked = uiState.isBeginnerViewLocked,
-                targetBallDistance = uiState.targetBallDistance,
+                targetBallDistance = uiState.targetBallDistance.toFloat(),
                 distanceUnit = uiState.distanceUnit,
                 onCycleTableSize = { onEvent(MainScreenEvent.CycleTableSize) },
                 onMenuClick = { onEvent(MainScreenEvent.ToggleNavigationRail) }
